# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GCDR API
  description: |
    API para o sistema GCDR (Global Customer Data Repository).
    Sistema multi-tenant para gerenciamento de clientes, assets, dispositivos, regras e integrações.
  version: 1.0.0
  contact:
    name: GCDR Team
  license:
    name: Proprietary
servers:
  - url: https://9gc49yiru7.execute-api.sa-east-1.amazonaws.com/dev
    description: Development
  - url: https://api.gcdr.io/staging
    description: Staging
  - url: https://api.gcdr.io
    description: Production
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    TenantId:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant identifier for multi-tenancy
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
      required:
        - success
        - data
...
    User:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        partnerId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        type:
          type: string
          enum: [INTERNAL, CUSTOMER, PARTNER, SERVICE_ACCOUNT]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED, PENDING_VERIFICATION, LOCKED]
        profile:
          $ref: '#/components/schemas/UserProfile'
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        mfaEnabled:
          type: boolean
        emailVerified:
          type: boolean
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
        phone:
          type: string
        department:
          type: string
        jobTitle:
          type: string
        bio:
          type: string
    UserPreferences:
      type: object
      properties:
        language:
          type: string
        timezone:
          type: string
        dateFormat:
          type: string
        timeFormat:
          type: string
          enum: ['12h', '24h']
        theme:
          type: string
          enum: [light, dark, system]
        notifications:
          $ref: '#/components/schemas/NotificationPreferences'
        dashboardLayout:
          type: object
    NotificationPreferences:
      type: object
      properties:
        email:
          type: boolean
        push:
          type: boolean
        sms:
          type: boolean
        inApp:
          type: boolean
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user
      operationId: getCurrentUser
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/User'
...